Update to use lubridate and set NA ranges
Use lubridate functions to get a slightly more accurate duration between birth date and current date. Also lubridate is part of expanded tidyverse so more pipe friendly, so changing this function to use piping as well. In terms of NA ranges for age, some birthdate data is clearly wrong, so changing code to first set NA for birth dates that are clearly wrong - those who indicate that they are under 18 or over 100 years of age based on birth year.